Backend/services/image_service/requirements.txt
fastapi==0.111.0
uvicorn==0.21.1
python-multipart==0.0.9
cloudinary==1.44.0
python-dotenv==1.0.0
pydantic==2.7.4
httpx==0.24.1 